@@ -13,7 +13,8 @@
     "javscript",
     "markdown",
     "notebook",
-    "mongodb"
+    "mongodb",
+    "kayero"
   ],
   "author": "Mathieu Dutour <mathieu.dutour@me.com>",
   "license": "MIT",
@@ -22,37 +23,37 @@
   },
   "homepage": "https://github.com/mathieudutour/kayero",
   "dependencies": {
-    "clipboard": "^1.5.10",
-    "electron-window-state": "3.0.3",
-    "katex": "0.6.0",
-    "markdown-it-katex": "2.0.1",
-    "monk": "3.0.7"
+    "clipboard": "^1.6.1",
+    "electron-debug": "1.1.0",
+    "electron-window-state": "4.1.1",
+    "monk": "4.0.0"
   },
   "devDependencies": {
-    "ava": "^0.15.2",
+    "ava": "^0.19.1",
     "babel-core": "^6.9.1",
-    "babel-eslint": "^6.0.4",
-    "babel-loader": "^6.2.4",
+    "babel-eslint": "^7.2.3",
+    "babel-loader": "^7.0.0",
     "babel-preset-es2015": "^6.6.0",
     "babel-preset-react": "^6.5.0",
     "babel-register": "^6.7.2",
-    "codecov": "^1.0.1",
+    "codecov": "^2.2.0",
     "codemirror": "^5.15.2",
     "coveralls": "^2.11.9",
     "css-loader": "^0.23.1",
-    "electron-debug": "^1.0.0",
-    "electron-osx-sign": "0.4.0-beta4",
-    "electron-packager": "^7.0.3",
-    "electron-prebuilt": "^1.2.2",
+    "devtron": "^1.4.0",
+    "electron": "^1.6.7",
+    "electron-osx-sign": "0.4.4",
+    "electron-packager": "^8.7.0",
     "electron-rebuild": "^1.1.5",
-    "eslint": "^2.12.0",
-    "eslint-config-standard": "^5.3.1",
-    "eslint-plugin-import": "^1.8.1",
-    "eslint-plugin-promise": "^1.3.2",
-    "eslint-plugin-react": "^5.1.1",
-    "eslint-plugin-standard": "^1.3.2",
-    "fetch-mock": "^4.4.0",
-    "file-loader": "^0.8.5",
+    "eslint": "^3.19.0",
+    "eslint-config-standard": "^10.2.1",
+    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-node": "4.2.2",
+    "eslint-plugin-promise": "^3.5.0",
+    "eslint-plugin-react": "^7.0.0",
+    "eslint-plugin-standard": "^3.0.1",
+    "fetch-mock": "^5.10.0",
+    "file-loader": "^0.11.1",
     "font-awesome": "^4.5.0",
     "front-matter": "^2.0.6",
     "highlight.js": "^9.2.0",
@@ -60,35 +61,43 @@
     "jsdom": "^9.2.1",
     "json-loader": "^0.5.4",
     "jutsu": "^0.1.1",
-    "markdown-it": "^6.0.1",
-    "mock-require": "^1.3.0",
-    "node-sass": "^3.7.0",
-    "nyc": "^6.6.1",
+    "lodash.debounce": "4.0.8",
+    "markdown-it": "^8.3.1",
+    "markdown-it-mathjax": "^2.0.0",
+    "mathjax": "^2.7.1",
+    "mock-require": "^2.0.2",
+    "node-sass": "^4.5.2",
+    "nyc": "^10.3.2",
+    "prop-types": "15.5.10",
     "query-string": "^4.1.0",
     "react": "^15.1.0",
-    "react-codemirror": "^0.2.6",
-    "react-dnd": "2.1.4",
-    "react-dnd-html5-backend": "2.1.2",
+    "react-codemirror": "^0.3.0",
+    "react-dnd": "^2.4.0",
+    "react-dnd-html5-backend": "^2.4.1",
     "react-dom": "^15.1.0",
     "react-hot-loader": "^1.3.0",
-    "react-redux": "^4.4.1",
+    "react-portal": "3.1.0",
+    "react-redux": "^5.0.4",
     "redux": "^3.3.1",
+    "redux-batched-actions": "^0.2.0",
+    "redux-logger": "^3.0.1",
     "redux-mock-store": "^1.0.2",
     "redux-thunk": "^2.0.1",
     "request": "^2.72.0",
-    "sass-loader": "^3.2.0",
-    "sinon": "^1.17.4",
+    "sass-loader": "^6.0.5",
+    "sinon": "^2.2.0",
     "sinon-as-promised": "^4.0.0",
     "smolder": "^0.3.1",
-    "style-loader": "^0.13.1",
-    "stylelint": "^6.5.1",
-    "stylelint-config-standard": "^8.0.0",
+    "string-hash": "1.1.3",
+    "style-loader": "^0.17.0",
+    "stylelint": "^7.10.1",
+    "stylelint-config-standard": "^16.0.0",
     "url-loader": "^0.5.7",
-    "webpack": "^1.13.1",
-    "webpack-dev-server": "^1.14.1",
+    "webpack": "^2.5.1",
+    "webpack-dev-server": "^2.4.5",
     "webpack-target-electron-renderer": "^0.4.0",
-    "whatwg-fetch": "^1.0.0",
-    "write-file-webpack-plugin": "^3.1.8"
+    "whatwg-fetch": "^2.0.3",
+    "write-file-webpack-plugin": "^4.0.2"
   },
   "scripts": {
     "test": "make test",
