@@ -1,66 +1,37 @@
 import React, { Component } from 'react'
+import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 import Title from './Title'
-import Metadata from './Metadata'
+import Dependencies from './Dependencies'
 import { metadataSelector } from '../selectors'
-import { toggleEdit, toggleSave, undo, fetchData } from '../actions'
+import { toggleSave } from '../actions'
 
 class Header extends Component {
-
   constructor (props) {
     super(props)
-    this.toggleEditClicked = this.toggleEditClicked.bind(this)
     this.toggleSaveClicked = this.toggleSaveClicked.bind(this)
-    this.undoClicked = this.undoClicked.bind(this)
-  }
-
-  toggleEditClicked () {
-    this.props.dispatch(toggleEdit())
   }
 
   toggleSaveClicked () {
     this.props.dispatch(toggleSave())
   }
 
-  undoClicked () {
-    this.props.dispatch(undo())
-    this.props.dispatch(fetchData())
-  }
-
   render () {
-    const { metadata, editable, undoSize, dispatch } = this.props
+    const { metadata, dispatch } = this.props
     const title = metadata.get('title')
-    const icon = editable ? 'fa-newspaper-o' : 'fa-pencil'
     document.title = title
-    const saveButton = (
-      <i className='fa fa-save' onClick={this.toggleSaveClicked}
-        title='Export notebook' />
-    )
-    const undoButton = (
-      <i className='fa fa-rotate-left' onClick={this.undoClicked} title='Undo' />
-    )
-    const changesMade = editable && undoSize > 0
     return (
       <div>
-        <Title title={title} editable={editable} dispatch={dispatch} />
-        <span className='controls'>
-          {changesMade ? undoButton : null}
-          {changesMade ? saveButton : null}
-          <i className={'fa ' + icon} onClick={this.toggleEditClicked}
-              title={editable ? 'Exit edit mode' : 'Enter edit mode'} />
-        </span>
-        <Metadata editable={editable} metadata={metadata} dispatch={dispatch} />
+        <Title title={title} dispatch={dispatch} />
+        <Dependencies dispatch={dispatch} datasources={metadata.get('datasources')} libraries={metadata.get('libraries')} />
       </div>
     )
   }
-
 }
 
 Header.propTypes = {
-  metadata: React.PropTypes.object,
-  editable: React.PropTypes.bool,
-  undoSize: React.PropTypes.number,
-  dispatch: React.PropTypes.func
+  metadata: PropTypes.object,
+  dispatch: PropTypes.func
 }
 
 export default connect(metadataSelector)(Header)
