@@ -1,4 +1,5 @@
 import React, { Component } from 'react'
+import PropTypes from 'prop-types'
 import { typeString, getSpacing } from './Visualiser'
 
 function buildCssClass (type, useHljs) {
@@ -10,6 +11,8 @@ function buildCssClass (type, useHljs) {
       cssSuffix = 'number'; break
     case 'Boolean':
       cssSuffix = 'literal'; break
+    case 'Error':
+      cssSuffix = 'error'; break
     case 'Function':
       cssSuffix = 'keyword'; break
     default:
@@ -23,12 +26,13 @@ function buildCssClass (type, useHljs) {
 }
 
 export default class DefaultVisualiser extends Component {
-
   render () {
     const { data, indent, name, useHljs, path, click = () => {} } = this.props
     const type = typeString(data)
-    const repr = (type === 'String') ? "'" + String(data) + "'"
-      : (type === 'Function') ? 'function()' : String(data)
+    const repr =
+      (type === 'String') ? "'" + String(data) + "'"
+      : (type === 'Function') ? 'function()'
+      : (type === 'Error') ? (data.stack || String(data)) : String(data)
     const cssClass = buildCssClass(type, useHljs)
     let key = <span className='visualiser-spacing'></span>
     if (name) {
@@ -53,20 +57,19 @@ export default class DefaultVisualiser extends Component {
       </div>
     )
   }
-
 }
 
 DefaultVisualiser.propTypes = {
-  data: React.PropTypes.oneOfType([
-    React.PropTypes.string,
-    React.PropTypes.number,
-    React.PropTypes.func,
-    React.PropTypes.instanceOf(Error),
-    React.PropTypes.instanceOf(Date)
+  data: PropTypes.oneOfType([
+    PropTypes.string,
+    PropTypes.number,
+    PropTypes.func,
+    PropTypes.instanceOf(Error),
+    PropTypes.instanceOf(Date)
   ]),
-  indent: React.PropTypes.number,
-  useHljs: React.PropTypes.string,
-  name: React.PropTypes.string,
-  path: React.PropTypes.string,
-  click: React.PropTypes.func
+  indent: PropTypes.number,
+  useHljs: PropTypes.string,
+  name: PropTypes.string,
+  path: PropTypes.string,
+  click: PropTypes.func
 }
