@@ -6,11 +6,11 @@ import Immutable from 'immutable'
 import Smolder from 'smolder'
 /* global d3, nv */
 import Jutsu from 'jutsu' // Imports d3 and nv as globals
-import {remote as electron} from 'electron' // eslint-disable-line
+import {remote as electron} from 'electron'
 
 import { arrayToCSV } from './util'
-import { gistUrl, gistApi } from './config' // eslint-disable-line
-import { render } from './markdown'
+import { gistApi } from './config'
+import { render, extractBlocks } from './markdown'
 
 /*
  * Action types
@@ -20,16 +20,15 @@ export const CODE_RUNNING = 'CODE_RUNNING'
 export const CODE_EXECUTED = 'CODE_EXECUTED'
 export const CODE_ERROR = 'CODE_ERROR'
 export const RECEIVED_DATA = 'RECEIVED_DATA'
-export const TOGGLE_EDIT = 'TOGGLE_EDIT'
-export const UPDATE_BLOCK = 'UPDATE_BLOCK'
-export const EDIT_BLOCK = 'EDIT_BLOCK'
+export const UPDATE_CONTENT = 'UPDATE_CONTENT'
 export const UPDATE_META = 'UPDATE_META'
 export const TOGGLE_META = 'TOGGLE_META'
 export const ADD_BLOCK = 'ADD_BLOCK'
 export const DELETE_BLOCK = 'DELETE_BLOCK'
-export const MOVE_BLOCK = 'MOVE_BLOCK'
 export const DELETE_DATASOURCE = 'DELETE_DATASOURCE'
 export const UPDATE_DATASOURCE = 'UPDATE_DATASOURCE'
+export const DELETE_LIBRARY = 'DELETE_LIBRARY'
+export const UPDATE_LIBRARY = 'UPDATE_LIBRARY'
 export const TOGGLE_SAVE = 'TOGGLE_SAVE'
 export const GIST_CREATED = 'GIST_CREATED'
 export const UNDO = 'UNDO'
@@ -103,14 +102,35 @@ export function openFile () {
       })
     }).then(() => dispatch(fetchData()))
   }
-};
+}
+
+export function updateContent (content) {
+  electron.BrowserWindow.getFocusedWindow().setDocumentEdited(true)
+  const {blocks, blockOrder} = extractBlocks(content)
+  return (dispatch, getState) => {
+    const oldBlockOrder = getState().notebook.get('blockOrder').toJS()
+    return Promise.resolve().then(() =>
+      dispatch({
+        type: UPDATE_CONTENT,
+        content,
+        blocks: Immutable.fromJS(blocks),
+        blockOrder: Immutable.fromJS(blockOrder)
+      })
+    ).then(() =>
+      Promise.all([
+        ...oldBlockOrder.filter(old => blockOrder.indexOf(old) === -1).map(id => deleteBlock(id)),
+        ...blockOrder.filter(newBlock => oldBlockOrder.indexOf(newBlock) === -1).map(id => addCodeBlock(id))
+      ])
+    )
+  }
+}
 
 function initDBs (getState) {
   const executionState = getState().execution
   let data = {}
   if (executionState) {
     const immutableData = executionState.get('data')
-    data = immutableData && immutableData.toJS() || {}
+    data = (immutableData && immutableData.toJS()) || {}
   }
   const notebook = getState().notebook
   let filePath
@@ -147,21 +167,43 @@ function initDBs (getState) {
 
 function closeDBs (dbs) {
   Object.keys(dbs).forEach((k) => dbs[k].close())
-  return
+}
+
+export function executeAuto () {
+  return (dispatch, getState) => {
+    const notebook = getState().notebook
+    const blocks = notebook.get('blocks')
+    const order = notebook.get('blockOrder')
+
+    // init the database connections
+    const dbs = initDBs(getState)
+
+    // This slightly scary Promise chaining ensures that code blocks are executed in order.
+    return order.reduce((p, id) => {
+      return p.then(() => {
+        const option = blocks.getIn([String(id), 'option'])
+        if (option === 'auto' || option === 'hidden') {
+          return dispatch(executeCodeBlock(String(id), dbs))
+        }
+        return Promise.resolve()
+      })
+    }, Promise.resolve()).then(() => closeDBs(dbs))
+  }
 }
 
 export function executeCodeBlock (id, dbs) {
   return (dispatch, getState) => {
     dispatch(codeRunning(id))
-    const code = getState().notebook.getIn(['blocks', id, 'content'])
+    const {notebook, execution} = getState()
+    const code = notebook.getIn(['blocks', id, 'content'])
     const graphElement = document.getElementById('kayero-graph-' + id)
 
-    const executionState = getState().execution
-    const context = executionState.get('executionContext').toJS()
-    const data = executionState.get('data').toJS()
+    const context = execution.get('executionContext').toJS()
+    const data = execution.get('data').toJS()
 
-    const soloExecution = !dbs
+    let shouldCloseDbs = false
     if (!dbs) {
+      shouldCloseDbs = true
       dbs = initDBs(getState)
     }
     Object.keys(dbs).forEach((k) => {
@@ -170,14 +212,19 @@ export function executeCodeBlock (id, dbs) {
       }
     })
 
+    const libs = notebook.getIn(['metadata', 'libraries']).reduce((prev, url, name) => {
+      prev[name] = window[name]
+      return prev
+    }, {})
+
     const jutsu = Smolder(Jutsu(graphElement))
 
     return new Promise((resolve, reject) => {
       try {
         const result = new Function(
-          ['d3', 'nv', 'graphs', 'data', 'graphElement'], code
+          ['d3', 'nv', 'graphs', 'data', 'graphElement', 'libs'], code
         ).call(
-          context, d3, nv, jutsu, data, graphElement
+          context, d3, nv, jutsu, data, graphElement, libs
         )
         resolve(result)
       } catch (err) {
@@ -192,7 +239,7 @@ export function executeCodeBlock (id, dbs) {
       dispatch(codeError(id, err))
     })
     .then(() => {
-      if (soloExecution) {
+      if (shouldCloseDbs) {
         return closeDBs(dbs)
       }
     })
@@ -223,35 +270,31 @@ function codeError (id, err) {
   }
 }
 
-export function executeAuto () {
-  return (dispatch, getState) => {
-    const notebook = getState().notebook
-    const blocks = notebook.get('blocks')
-    const order = notebook.get('content')
-
-    // init the database connections
-    const dbs = initDBs(getState)
-
-    // This slightly scary Promise chaining ensures that code blocks
-    // are executed in order, even if they return Promises.
-    return order.reduce((p, id) => {
-      return p.then(() => {
-        const option = blocks.getIn([id, 'option'])
-        if (option === 'auto' || option === 'hidden') {
-          return dispatch(executeCodeBlock(id, dbs))
-        }
-        return Promise.resolve()
-      })
-    }, Promise.resolve()).then(() => closeDBs(dbs))
-  }
-}
-
 function receivedData (name, data) {
   return {
     type: RECEIVED_DATA,
     name,
     data
   }
+}
+
+function importScript (url) {
+  return new Promise((resolve, reject) => {
+    var scripts = document.getElementsByTagName('script')
+    for (var i = scripts.length; i--;) {
+      if (scripts[i].src === url) { // check if already imported
+        resolve()
+        return
+      }
+    }
+    // otherwise import it
+    const oScript = document.createElement('script')
+    oScript.type = 'text/javascript'
+    oScript.onerror = reject
+    oScript.onload = resolve
+    oScript.src = url
+    document.head.appendChild(oScript)
+  })
 }
 
 export function fetchData () {
@@ -285,25 +328,15 @@ export function fetchData () {
         }
       }
     )
+
+    getState().notebook.getIn(['metadata', 'libraries']).forEach((url) => {
+      proms.push(importScript(url))
+    })
+
     // When all data fetched, run all the auto-running code blocks.
     return Promise.all(proms).then(() => dispatch(executeAuto()))
   }
 }
-
-export function toggleEdit () {
-  return {
-    type: TOGGLE_EDIT
-  }
-}
-
-export function updateBlock (id, text) {
-  electron.BrowserWindow.getFocusedWindow().setDocumentEdited(true)
-  return {
-    type: UPDATE_BLOCK,
-    id,
-    text
-  }
-};
 
 export function updateTitle (text) {
   return {
@@ -329,26 +362,19 @@ export function toggleFooter () {
 };
 
 export function addCodeBlock (id) {
-  return {
-    type: ADD_BLOCK,
-    blockType: 'code',
-    id
-  }
-};
-
-export function addTextBlock (id) {
-  return {
-    type: ADD_BLOCK,
-    blockType: 'text',
-    id
-  }
-};
-
-export function addGraphBlock (id) {
-  return {
-    type: ADD_BLOCK,
-    blockType: 'graph',
-    id
+  return (dispatch, getState) => {
+    return Promise.resolve().then(() =>
+      dispatch({
+        type: ADD_BLOCK,
+        blockType: 'code',
+        id
+      })
+    ).then(() => {
+      const option = getState().notebook.getIn(['blocks', id, 'option'])
+      if (option === 'auto') {
+        return dispatch(executeCodeBlock(String(id)))
+      }
+    })
   }
 };
 
@@ -359,14 +385,6 @@ export function deleteBlock (id) {
   }
 };
 
-export function moveBlock (id, nextIndex) {
-  return {
-    type: MOVE_BLOCK,
-    id,
-    nextIndex
-  }
-}
-
 export function deleteDatasource (id) {
   return {
     type: DELETE_DATASOURCE,
@@ -375,10 +393,37 @@ export function deleteDatasource (id) {
 };
 
 export function updateDatasource (id, url) {
+  return (dispatch, getState) => {
+    return Promise.resolve().then(() =>
+      dispatch({
+        type: UPDATE_DATASOURCE,
+        id,
+        text: url
+      })
+    ).then(() =>
+      dispatch(fetchData())
+    )
+  }
+};
+
+export function deleteLibrary (id) {
   return {
-    type: UPDATE_DATASOURCE,
-    id,
-    text: url
+    type: DELETE_DATASOURCE,
+    id
+  }
+};
+
+export function updateLibrary (id, url) {
+  return (dispatch, getState) => {
+    return Promise.resolve().then(() =>
+      dispatch({
+        type: UPDATE_DATASOURCE,
+        id,
+        text: url
+      })
+    ).then(() =>
+      dispatch(fetchData())
+    )
   }
 };
 
@@ -554,13 +599,6 @@ export function compileGraphBlock (id) {
 export function clearGraphData (id) {
   return {
     type: CLEAR_GRAPH_BLOCK_DATA,
-    id
-  }
-}
-
-export function editBlock (id) {
-  return {
-    type: EDIT_BLOCK,
     id
   }
 }
